requirements.txt
alembic==1.15.2
cryptography==44.0.2
fastapi==0.115.11
httpx==0.28.1
itsdangerous==2.2.0
jinja2==3.1.6
pydantic-settings==2.8.1
pymysql==1.1.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-multipart==0.0.20
sqlalchemy==2.0.39
uvicorn==0.34.0
